🔨 Extract test to a separate folder 🔨
